#414: refactor: add commonly used test utils
