+++ tendermint/tests/utils/lite.rs
@@ -9,8 +9,6 @@ use tendermint::lite::Requester;
 use tendermint::{
     block::signed_header::SignedHeader, evidence::Duration, lite, validator::Set, Time,
 };
-use std::{process, io};
-use std::io::Read;
 
 /// Test that a struct `T` can be:
 ///
@@ -18,8 +16,8 @@ use std::io::Read;
 /// - parsed back from the serialized JSON of the previous step
 /// - that the two parsed structs are equal according to their `PartialEq` impl
 pub fn test_serialization_roundtrip<T>(obj: &T)
-where
-    T: Debug + PartialEq + Serialize + DeserializeOwned,
+    where
+        T: Debug + PartialEq + Serialize + DeserializeOwned,
 {
     let serialized = serde_json::to_string(obj).unwrap();
     let parsed = serde_json::from_str(&serialized).unwrap();
@@ -91,84 +89,4 @@ impl MockRequester {
             validators: val_map,
         }
     }
-}
-
-/// A thin wrapper around process::Command to facilitate running external commands.
-pub struct Command {
-    program: Option<String>,
-    args: Vec<String>,
-    dir: Option<String>
-}
-
-impl Command {
-    /// Constructs a new Command for the given program with arguments.
-    pub fn new() -> Command {
-        Command {
-            program: None,
-            args: vec![],
-            dir: None
-        }
-    }
-
-    /// Sets the program to run
-    pub fn program(&mut self, program: &str) -> &mut Self {
-        self.program = Some(program.to_owned());
-        self
-    }
-
-    /// Adds a new program argument
-    pub fn arg(&mut self, arg: &str) -> &mut Self {
-        self.args.push(arg.to_owned());
-        self
-    }
-
-    /// Adds a new program argument, concatenated from several parts
-    pub fn arg_from_parts(&mut self, parts: Vec<&str>) -> &mut Self {
-        let mut arg: String = String::new();
-        for part in parts {
-            arg = arg + part;
-        }
-        self.args.push(arg);
-        self
-    }
-
-    /// Sets the working directory for the child process
-    pub fn current_dir(&mut self, dir: &str) -> &mut Self {
-        self.dir = Some(dir.to_owned());
-        self
-    }
-
-    /// Executes the command as a child process, and extracts its status, stdout, stderr.
-    pub fn spawn(&mut self) -> io::Result<CommandRun> {
-        match &self.program {
-            None => Err(io::Error::new(io::ErrorKind::InvalidInput, "")),
-            Some(program) => {
-                let mut command = process::Command::new(program);
-                command.args(&self.args)
-                    .stdout(process::Stdio::piped())
-                    .stderr(process::Stdio::piped());
-                if let Some(dir) = &self.dir {
-                    command.current_dir(dir);
-                }
-                let mut process = command.spawn()?;
-                let status = process.wait()?;
-                let mut stdout = String::new();
-                process.stdout.unwrap().read_to_string(&mut stdout)?;
-                let mut stderr = String::new();
-                process.stderr.unwrap().read_to_string(&mut stderr)?;
-                Ok(CommandRun {
-                    status,
-                    stdout,
-                    stderr
-                })
-            }
-        }
-    }
-}
-
-/// The result of a command execution if managed to run the child process
-pub struct CommandRun {
-    pub status: process::ExitStatus,
-    pub stdout: String,
-    pub stderr: String
 }
